% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(num1,num2)
}
\arguments{
\item{num1}{an integer number}

\item{num2}{an integer number}
}
\value{
GCD
}
\description{
this function recieves two numbers and returns GCD.
}
\details{
euclidean which calculates Greater Common Divisor of two numbers.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\author{
Hamed
}
